change lcd timer method to int

+++ src/lcd/nextion.cpp
@@ -239,7 +239,7 @@ void lcdBrewTimerStop(void) {
   myNex.writeNum("timerState", 0);
 }
 
-void lcdSetBrewTimer(long seconds) {
+void lcdSetBrewTimer(int seconds) {
   myNex.writeNum("activeBrewTime", seconds);
 }
